+++ Cargo.toml
@@ -9,6 +9,7 @@ edition = "2018"
 [dependencies]
 parking_lot = "0.10.2"
 once_cell = "1.3.1"
+futures-io = "0.3.5"
 
 [dependencies.iou]
 git = "https://github.com/withoutboats/iou"
